User registration updated

+++ src/main/java/org/example/controller/RoleController.java
@@ -17,7 +17,7 @@ public class RoleController {
     }
 
     @PutMapping("/update/{id}")
-    public String updateRole(@PathVariable Long id, @RequestBody RoleDto roleDto){
+    public String updateRole(@PathVariable Short id, @RequestBody RoleDto roleDto){
         roleDto.setId(id);
         if(service.updateRole(roleDto)){
             return "Updated";
@@ -26,7 +26,7 @@ public class RoleController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteRole(@PathVariable Long id) {
+    public String deleteRole(@PathVariable Short id) {
         if (service.deleteRole(id)) {
             return "Deleted";
         }
@@ -39,7 +39,7 @@ public class RoleController {
     }
 
     @GetMapping("/search-by-id/{id}")
-    public RoleDto searchUserById(@PathVariable Long id){
+    public RoleDto searchUserById(@PathVariable Short id){
         return service.searchRoleById(id);
     }
 }
