+++ src/main/java/org/example/repository/RoleRepository.java
@@ -3,5 +3,5 @@ package org.example.repository;
 import org.example.entity.RoleEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RoleRepository extends CrudRepository<RoleEntity,Long> {
+public interface RoleRepository extends CrudRepository<RoleEntity,Short> {
 }
